prihlaseny pouzivatel edit

+++ Skladnik/src/main/java/dao/MaterialDao.java
@@ -19,6 +19,8 @@ public interface MaterialDao {
     void odstranMaterial(Material material);
     public List <Material> getAll();
     List<String> dajNazovVsetkychMaterialov();
-    void upravStavMaterial(long idMaterialu, long pocet,long typPohybu);
+    void upravStavMaterial(long idMaterialu, double pocet, long typPohybu);
     void upravCenuMaterialu(long idMaterialu, double cena);
+
+    
 }
